Address PR review: move RHAIIS-specific code out of core/caliper
- Move PROFILE_MAP, METRICS from caliper engine to rhaiis postprocess
- Make compare_runs/run_regression_analysis accept profile_map and
  metrics as parameters instead of hardcoded constants
- Move send_regression_notification, dashboard URL builder, and profile
  display names from core/notifications to rhaiis postprocess
- Consolidate Slack webhook into psap-forge-notifications vault,
  remove separate psap-forge-rhaiis-slack vault
- Remove hardcoded DEFAULT_AGENT_URL from agent.py, require caller
  to provide URL from config
- Add docstring to enable_profiler_gate run function
- Clean up vault descriptions to be project-generic
