+++ projects/rhaiis/orchestration/config.d/rhaiis.yaml
@@ -4,7 +4,7 @@ cluster_tag: ""
 
 gpu_types:
   nvidia: h200
-  amd_mi300x: mi300x
+  amd: mi300x
   l40s: l40s
 
 images:
@@ -17,7 +17,7 @@ deploy:
   memory_request: "16Gi"
   storage_source: hf
   storage_pvc: model-pvc
-  image_pull_secret: npalaska-image-pull
+  image_pull_secret: ""
   service_account_name: ""
   ready_timeout: 3600
   health_check_timeout: 120
